fix email handler in ContactForm.astro
