+++ package.json
@@ -10,38 +10,39 @@
     "lint": "astro check"
   },
   "dependencies": {
-    "astro": "^5.16.11",
-    "@astrojs/react": "^4.4.2",
-    "@astrojs/tailwind": "^6.0.2",
-    "@astrojs/sitemap": "^3.7.0",
     "@astrojs/node": "^9.5.2",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "tailwindcss": "^3.4.19",
-    "class-variance-authority": "^0.7.1",
-    "clsx": "^2.1.1",
-    "tailwind-merge": "^3.4.0",
-    "lucide-react": "^0.562.0",
-    "@radix-ui/react-slot": "^1.2.4",
+    "@astrojs/react": "^4.4.2",
+    "@astrojs/sitemap": "^3.7.0",
+    "@astrojs/tailwind": "^6.0.2",
     "@radix-ui/react-dialog": "^1.1.15",
     "@radix-ui/react-dropdown-menu": "^2.1.16",
     "@radix-ui/react-label": "^2.1.8",
     "@radix-ui/react-select": "^2.2.6",
     "@radix-ui/react-separator": "^1.1.8",
+    "@radix-ui/react-slot": "^1.2.4",
+    "astro": "^5.16.11",
+    "class-variance-authority": "^0.7.1",
+    "clsx": "^2.1.1",
     "formik": "^2.2.9",
-    "yup": "^0.32.11",
     "load-script": "^2.0.0",
+    "lucide-react": "^0.562.0",
+    "mailgun-js": "^0.22.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-icons": "^4.4.0",
     "sass": "^1.54.8",
-    "mailgun-js": "^0.22.0"
+    "tailwind-merge": "^3.4.0",
+    "tailwindcss": "^3.4.19",
+    "yup": "^0.32.11"
   },
   "devDependencies": {
+    "@tailwindcss/typography": "^0.5.19",
+    "@types/mailgun-js": "^0.22.19",
     "@types/react": "^18.0.18",
     "@types/react-dom": "^18.0.18",
-    "typescript": "^5.0.0",
     "autoprefixer": "^10.4.23",
     "eslint": "^8.23.0",
     "prettier": "^2.7.1",
-    "@tailwindcss/typography": "^0.5.19"
+    "typescript": "^5.0.0"
   }
 }
